[RUI-236]: Extract scroll logic to separate hook
