+++ src/components/charts/lines/LineChartTabbedPro/components/MeasureTabs.tsx
@@ -1,8 +1,9 @@
-import { FC, useCallback, useEffect, useRef, useState } from 'react';
+import { FC } from 'react';
 import { DataResponse, Measure } from '@embeddable.com/core';
 import { IconChevronLeft, IconChevronRight } from '@tabler/icons-react';
 import { Theme } from '../../../../../theme/theme.types';
 import { MeasureTab } from './MeasureTab';
+import { useHorizontalScroll } from './useHorizontalScroll';
 import styles from './MeasureTabs.module.css';
 
 type MeasureTabsProps = {
@@ -20,37 +21,8 @@ export const MeasureTabs: FC<MeasureTabsProps> = ({
   onTabClick,
   theme,
 }) => {
-  const scrollRef = useRef<HTMLDivElement>(null);
-  const [canScrollRight, setCanScrollRight] = useState(false);
-  const [canScrollLeft, setCanScrollLeft] = useState(false);
-
-  const updateScrollState = useCallback(() => {
-    const el = scrollRef.current;
-    if (!el) return;
-    setCanScrollLeft(el.scrollLeft > 1);
-    setCanScrollRight(el.scrollLeft + el.clientWidth < el.scrollWidth - 1);
-  }, []);
-
-  const handleScrollRight = () => {
-    scrollRef.current?.scrollBy({ left: 200, behavior: 'smooth' });
-  };
-
-  const handleScrollLeft = () => {
-    scrollRef.current?.scrollBy({ left: -200, behavior: 'smooth' });
-  };
-
-  useEffect(() => {
-    const el = scrollRef.current;
-    if (!el) return;
-    const ro = new ResizeObserver(updateScrollState);
-    ro.observe(el);
-    el.addEventListener('scroll', updateScrollState);
-    updateScrollState();
-    return () => {
-      ro.disconnect();
-      el.removeEventListener('scroll', updateScrollState);
-    };
-  }, [updateScrollState, measures]);
+  const { scrollRef, canScrollLeft, canScrollRight, handleScrollLeft, handleScrollRight } =
+    useHorizontalScroll([measures]);
 
   const totalsRow = resultsTotals?.data?.[0];
   const isLoading = !resultsTotals || resultsTotals.isLoading;
