+++ src/components/charts/lines/LineChartTabbedPro/components/useHorizontalScroll.ts
@@ -1,0 +1,37 @@
+import { useCallback, useEffect, useRef, useState, DependencyList } from 'react';
+
+export const useHorizontalScroll = (deps: DependencyList = []) => {
+  const scrollRef = useRef<HTMLDivElement>(null);
+  const [canScrollRight, setCanScrollRight] = useState(false);
+  const [canScrollLeft, setCanScrollLeft] = useState(false);
+
+  const updateScrollState = useCallback(() => {
+    const el = scrollRef.current;
+    if (!el) return;
+    setCanScrollLeft(el.scrollLeft > 1);
+    setCanScrollRight(el.scrollLeft + el.clientWidth < el.scrollWidth - 1);
+  }, []);
+
+  const handleScrollRight = () => {
+    scrollRef.current?.scrollBy({ left: 200, behavior: 'smooth' });
+  };
+
+  const handleScrollLeft = () => {
+    scrollRef.current?.scrollBy({ left: -200, behavior: 'smooth' });
+  };
+
+  useEffect(() => {
+    const el = scrollRef.current;
+    if (!el) return;
+    const ro = new ResizeObserver(updateScrollState);
+    ro.observe(el);
+    el.addEventListener('scroll', updateScrollState);
+    updateScrollState();
+    return () => {
+      ro.disconnect();
+      el.removeEventListener('scroll', updateScrollState);
+    };
+  }, [updateScrollState, ...deps]);
+
+  return { scrollRef, canScrollLeft, canScrollRight, handleScrollLeft, handleScrollRight };
+};
